+++ Plugin/CleverTapUnity/CleverTapUnity-Scripts/CleverTapBinding.cs
@@ -11,7 +11,7 @@ using CleverTap.Utilities;
 namespace CleverTap {
   public class CleverTapBinding : MonoBehaviour {
       
-    public const string Version = "2.4.0";
+  public const string Version = "2.4.0";
 
 #if UNITY_IOS
     void Start() {
@@ -152,6 +152,9 @@ namespace CleverTap {
 
     [System.Runtime.InteropServices.DllImport("__Internal")]
     private static extern void CleverTap_showAppInbox(string styleConfig);
+
+    [System.Runtime.InteropServices.DllImport("__Internal")]
+    private static extern void CleverTap_dismissAppInbox();
 
     [System.Runtime.InteropServices.DllImport("__Internal")]
     private static extern int CleverTap_getInboxMessageCount();
@@ -453,6 +456,10 @@ namespace CleverTap {
         CleverTap_showAppInbox(styleConfigString);
     }
 
+    public static void DismissAppInbox() {
+        CleverTap_dismissAppInbox();
+    }
+
     public static int GetInboxMessageCount() {
         return CleverTap_getInboxMessageCount();
     }
@@ -676,17 +683,17 @@ namespace CleverTap {
 
     public static void CreateNotificationChannelWithSound(string channelId,string channelName, string channelDescription, int importance, bool showBadge, string sound){
         AndroidJavaObject context = unityCurrentActivity.Call<AndroidJavaObject>("getApplicationContext");
-        CleverTapAPI.CallStatic("createNotificationChannel",context,channelId,channelName,channelDescription,importance,showBadge,sound);
+        CleverTapAPI.CallStatic("createNotificationChannelWithSound",context,channelId,channelName,channelDescription,importance,showBadge,sound);
     }
 
     public static void CreateNotificationChannelWithGroup(string channelId,string channelName, string channelDescription, int importance, string groupId, bool showBadge){
         AndroidJavaObject context = unityCurrentActivity.Call<AndroidJavaObject>("getApplicationContext");
-        CleverTapAPI.CallStatic("createNotificationChannelwithGroup",context,channelId,channelName,channelDescription,importance,groupId,showBadge);
+        CleverTapAPI.CallStatic("createNotificationChannelWithGroup",context,channelId,channelName,channelDescription,importance,groupId,showBadge);
     }
 
     public static void CreateNotificationChannelWithGroupAndSound(string channelId,string channelName, string channelDescription, int importance, string groupId, bool showBadge, string sound){
         AndroidJavaObject context = unityCurrentActivity.Call<AndroidJavaObject>("getApplicationContext");
-        CleverTapAPI.CallStatic("createNotificationChannelwithGroup",context,channelId,channelName,channelDescription,importance,groupId,showBadge,sound);
+        CleverTapAPI.CallStatic("createNotificationChannelWithGroupAndSound",context,channelId,channelName,channelDescription,importance,groupId,showBadge,sound);
     }
 
     public static void CreateNotificationChannelGroup(string groupId, string groupName){
@@ -953,6 +960,10 @@ namespace CleverTap {
          CleverTap.Call("showAppInbox", styleConfig);
     }
 
+    public static void DismissAppInbox(){
+         CleverTap.Call("dismissAppInbox");
+    }
+
     public static int GetInboxMessageCount(){
         return CleverTap.Call<int>("getInboxMessageCount");
     }
@@ -1154,6 +1165,9 @@ namespace CleverTap {
     }
 
     public static void ShowAppInbox(string styleConfig){
+    }
+
+    public static void DismissAppInbox(){
     }
 
     public static int GetInboxMessageCount(){