+++ Plugin/CleverTapUnity/CleverTapUnity-Scripts/CleverTapUnity.cs
@@ -231,7 +231,7 @@ public class CleverTapUnity: MonoBehaviour {
         Debug.Log("unity received inbox message button with custom extras select: " + (!String.IsNullOrEmpty(message) ? message : "NULL"));
     }
 
-    //returns data associated with inbox message item click
+    // returns on the click of app inbox message with a string of the inbox payload along with page index and button index
     void CleverTapInboxItemClicked(string message)
     {
         Debug.Log("unity received inbox message clicked callback: " + (!String.IsNullOrEmpty(message) ? message : "NULL"));